Create jog_buttons
I hope this really works. I couldn't really debug it so you'll have to finish it up with the real CNC table saw. Very important to check the directions in dict_of_buttons are correct, if it all runs backwards it probablly that. So glad I could be part of this amazing project!